DP-12820 followup - remove unused connect access and secret keys
## Background
This is a followup to DP-12820 to remove the connect_aws_credentials pending eventual deactivation as the secret is used in connect system tests
and contains a script to authenticate with the caas-dev 037803949979:semaphoreci user. This will be replaced with dynamic iam role based access.

## Changes
This PR removes the `connect_aws_credentials` which is likely not needed in your pipeline.

## Actions
If the PR build succeeds, and the secret is only used in the semaphore.yml file, please merge this change.

For questions, please reach out to #devprod-oncall, and for more details, please see: https://confluentinc.atlassian.net/wiki/spaces/TOOLS/pages/3033867557/CI+System+Migration+Jenkins-+Semaphore#Secrets